add bootstrap test
