+++ cmd/main.go
@@ -7,17 +7,11 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/ChainSafe/fil-secondary-retrieval-markets/client"
-	"github.com/ChainSafe/fil-secondary-retrieval-markets/network"
-	"github.com/ChainSafe/fil-secondary-retrieval-markets/shared"
 	"github.com/ipfs/go-cid"
 	logging "github.com/ipfs/go-log/v2"
-	libp2p "github.com/libp2p/go-libp2p"
-	"github.com/libp2p/go-libp2p-core/host"
-	peer "github.com/libp2p/go-libp2p-core/peer"
 	"github.com/urfave/cli"
 )
 
@@ -107,37 +101,4 @@ func run(ctx *cli.Context) error {
 			return nil
 		}
 	}
-}
-
-func newNetwork(bootnodesStr string) (*network.Network, error) {
-	ctx := context.Background()
-	h, err := libp2p.New(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	// bootstrap to network
-	strs := strings.Split(bootnodesStr, ",")
-	addrs, err := shared.StringsToAddrInfos(strs)
-	if err != nil {
-		return nil, err
-	}
-
-	err = bootstrap(h, addrs)
-	if err != nil {
-		return nil, err
-	}
-
-	return network.NewNetwork(h)
-}
-
-func bootstrap(h host.Host, bns []peer.AddrInfo) error {
-	ctx := context.Background()
-	for _, bn := range bns {
-		err := h.Connect(ctx, bn)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }
